requirements.txt
apache-beam[gcp]>=2.50.0
xarray>=2023.1.0
xarray-beam>=0.7.0
zarr>=2.16.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
dask>=2023.1.0
matplotlib>=3.7.0
cartopy>=0.21.0
pyproj>=3.4.0
anthropic>=0.34.0
python-dotenv>=1.0.0